Support embedding by importing AwaEmbedding
